fix: remove /teachers route and redirect to /
